misc: refactor and add "clear compile log"

+++ app/src/main/java/mod/hosni/fraj/compilerlog/CompileErrorSaver.java
@@ -1,24 +1,13 @@
 package mod.hosni.fraj.compilerlog;
 
-import static mod.SketchwareUtil.getDip;
-
 import android.app.AlertDialog;
 import android.content.Context;
-import android.content.DialogInterface;
 import android.content.Intent;
-import android.graphics.Typeface;
-import android.view.View;
-import android.widget.HorizontalScrollView;
-import android.widget.ScrollView;
-import android.widget.TextView;
 
 import com.besome.sketch.tools.CompileLogActivity;
-import com.sketchware.remod.Resources;
 
-import mod.SketchwareUtil;
 import mod.agus.jcoderz.lib.FilePathUtil;
 import mod.agus.jcoderz.lib.FileUtil;
-import mod.hey.studios.util.CompileLogHelper;
 
 public class CompileErrorSaver {
 
@@ -36,7 +25,6 @@ public class CompileErrorSaver {
     public CompileErrorSaver(String sc_id) {
         this.sc_id = sc_id;
         path = FilePathUtil.getLastCompileLogPath(sc_id);
-        check();
     }
 
     /**
@@ -44,8 +32,8 @@ public class CompileErrorSaver {
      *
      * @param errorText The text to save, if possible, with detailed messages
      */
-    public void setErrorText(String errorText) {
-        FileUtil.deleteFile(path);
+    public void writeLogsToFile(String errorText) {
+        if (logFileExists()) FileUtil.deleteFile(path);
         FileUtil.writeFile(path, errorText);
     }
 
@@ -56,7 +44,8 @@ public class CompileErrorSaver {
      */
     public void showLastErrors(Context context) {
         Intent intent = new Intent(context, CompileLogActivity.class);
-        intent.putExtra("error", getLog());
+        intent.putExtra("error", getLogsFromFile());
+        intent.putExtra("sc_id", sc_id);
         intent.putExtra("showingLastError", true);
         intent.setFlags(Intent.FLAG_ACTIVITY_SINGLE_TOP);
         context.startActivity(intent);
@@ -65,24 +54,22 @@ public class CompileErrorSaver {
     /**
      * Clear the last saved error text.
      */
-    public void clear() {
+    public void deleteSavedLogs() {
         FileUtil.deleteFile(path);
     }
 
     /**
      * @return The last saved error text
      */
-    public String getLog() {
+    public String getLogsFromFile() {
+        if (!logFileExists()) return MESSAGE_NO_COMPILE_ERRORS_SAVED;
         return FileUtil.readFile(path);
     }
 
     /**
-     * Check if the last saved error text file exists, if not, it'll get created and
-     * {@link CompileErrorSaver#MESSAGE_NO_COMPILE_ERRORS_SAVED} will be written to it.
+     * Check if the last saved error text file exists.
      */
-    public void check() {
-        if (!FileUtil.isExistFile(path)) {
-            FileUtil.writeFile(path, MESSAGE_NO_COMPILE_ERRORS_SAVED);
-        }
+    public boolean logFileExists() {
+        return FileUtil.isExistFile(path);
     }
 }
